# PoS smaller rmap