`rmap` is a key data structure for finding matches. Unfortunately, it was quite large:
> PARAM_BC * size_of::<[Position; 2]>() = 15113 * 8 = 120904 bytes

While not a huge issue for CPU, it is still large enough to not fit into L1 CPU cache. The situation is even worse on GPUs, where it is extremely desirable to keep hot data in shared memory, which is also very small (~32 kiB on older GPUs, could be 64 kiB or larger on latest GPUs).

This PR exploits the fact that while the number of slots in `rmap` is 15113, the actual number of elements is upper-bound by a constant 272. So we can create two maps: one small map with 15113 small 16-bit pointers (I tried making it even smaller with 9-bit pointers, but on CPU it becomes slightly slower than larger `rmap` rather than faster, might be different on GPU though) into the second map with just 272 larger `[Position; 2]` elements.

Two maps require strictly more instructions, but benefit from smaller footprint:
> PARAM_BC * size_of::<u16>() + REDUCED_BUCKETS_SIZE * size_of::<[Position; 2]>() = 15113 * 2 + 272 * 8 = 30226 + 2176 = 32402 bytes

This fits nicely into 32 kiB limit of shared memory on all Vulkan-capable GPUs and ends up being a bit faster for multi-threaded CPU implementation too:
```
Before:
chia/table/parallel/8x  time:   [529.15 ms 534.42 ms 540.15 ms]
                        thrpt:  [14.811  elem/s 14.969  elem/s 15.119  elem/s]
After:
chia/table/parallel/8x  time:   [518.48 ms 521.57 ms 525.13 ms]
                        thrpt:  [15.234  elem/s 15.338  elem/s 15.430  elem/s]
```

If using 9-bit pointers will have adequate performance, it should be possible to fit both `rmap` and matches found in shared memory of 32 kiB, which would be very nice!